requirements.txt
fastapi
uvicorn
python-dotenv

langchain
langchain-community
langchain-openai
langchain-text-splitters

sentence-transformers
faiss-cpu
rank-bm25

transformers
torch

ragas
datasets

pypdf

numpy
pandas
scikit-learn